move hash utils out of hicache_storage to break CUDA import chain
get_hash_str and hash_str_to_int64 are pure-Python hash functions that
don't need CUDA. Moving them to a lightweight hash_utils module breaks
the import chain: radix_cache → hicache_storage → memory_pool_host →
sgl_kernel → libcuda.so.1. This allows io_struct.py (via schedule_batch
→ radix_cache) to be imported in CPU-only environments.
