+++ python/sglang/srt/managers/cache_controller.py
@@ -430,7 +430,7 @@ class HiCacheController:
         # Rollback-safe init: if creation fails, keep controller state consistent
         # for future attach attempts.
         self.storage_backend_type = storage_backend
-        from sglang.srt.mem_cache.hicache_storage import get_hash_str
+        from sglang.srt.mem_cache.hash_utils import get_hash_str
 
         self.get_hash_str = get_hash_str
         self.storage_config = self._generate_storage_config(
